various fixes and tweaks for monolithic build: removing duplicate defines of some structures, cleaning up some missing definitions a bit (headerfiles which turn out to be needed after all), etc.
Part of this work:

------------------

updating submodule MSVC2019 projects: libqrencode.

NOTE/WARNING: this is the odd one out as it has quite non-standard behaviour: it needs to **recompile the entire library source code with WITH_TESTS #defined** to ensure the various internal functions subjected to tests are indeed exposed and thus will be able to be linked against. (This is WIP (WorkInProgress) at qrencode-test project...)

As such, it (qrencode-test.vcxproj) MUST NOT HAVE A DEPENDENCY/REFERENCE to the library itself (libqrencode.vcxproj) as this would lead to double definitions and mixing slightly different codebases: a disaster waiting to happen.